+++ alerts/apps/frontend/src/components/AlertDetail/AlertReferencesRoutes/index.tsx
