+++ alerts/apps/frontend/src/components/AlertDetail/AlertReferencesRoutes/styles.module.css
