+++ alerts/apps/frontend/src/components/AlertDetail/AlertReferencesStops/index.tsx
