+++ alerts/apps/frontend/src/components/AlertDetail/AlertSectionCauseEffect/index.tsx
