+++ alerts/apps/frontend/src/components/AlertDetail/AlertSectionCauseEffect/styles.module.css
