+++ alerts/apps/frontend/src/components/AlertDetail/AlertSectionReferences/index.tsx
