+++ alerts/apps/frontend/src/components/AlertDetail/AlertSectionReferences/styles.module.css
