+++ alerts/apps/frontend/src/components/AlertDetail/AlertSectionTitle/index.tsx
