+++ alerts/apps/frontend/src/components/AlertDetail/AlertSectionValidity/index.tsx
