+++ alerts/apps/frontend/src/components/AlertDetail/AlertSectionValidity/styles.module.css
