+++ alerts/apps/frontend/src/components/AlertDetail/AlertSectionVisibility/index.tsx
