+++ alerts/apps/frontend/src/components/AlertDetail/AlertSectionVisibility/styles.module.css
